Updated the contact statement
Lots of people seem to be contacting me with basic questions that really
ought to be solved by reading the documentation. I'm happy to help out
if people want to contribute, or point out bugs I've introduced, but
people aren't even trying to use the mainline version before trying this
patchset. Oh well, the lure of git and colorized emails I guess!